:art: separate formatting functions from user detault component
