+++ src/screens/user-detail/UserDetail.tsx
@@ -6,10 +6,13 @@ import {RootStackParamsList} from '../Routes';
 import {LoadingIndicator} from '@src/components/loading-indicator/LoadingIndicator';
 
 import {Alert, SafeAreaView, Text, View} from 'react-native';
+
 import {globalStyles} from '@src/globalStyles';
+import {userDetailStyles} from './styles';
 
 import {gql, useQuery} from '@apollo/client';
-import {userDetailStyles} from './styles';
+
+import {formatDate, formatPhoneNumber} from '@src/utils/formatting';
 
 type UserDetailScreenProps = NativeStackScreenProps<
   RootStackParamsList,
@@ -76,17 +79,6 @@ export const UserDetail = ({route}: UserDetailScreenProps) => {
     );
   }
 
-  function formatPhoneNumber(phone: string) {
-    const ddd = phone.slice(0, 2);
-    const firstPart = phone.slice(2, 7);
-    const secondPart = phone.slice(7, 11);
-    return `(${ddd}) ${firstPart}-${secondPart}`;
-  }
-
-  function formatDate(dateString: string) {
-    return new Date(`${dateString}T00:00:00`).toLocaleDateString('pt-BR');
-  }
-
   const roleLabels = {
     admin: 'Administrador',
     user: 'Usuário',
@@ -105,24 +97,29 @@ export const UserDetail = ({route}: UserDetailScreenProps) => {
           <Text>
             <Text style={userDetailStyles.propertyTitle}>ID:</Text> {userId}
           </Text>
+
           <Text>
             <Text style={userDetailStyles.propertyTitle}>Nome:</Text>{' '}
             {user?.name}
           </Text>
+
           <Text>
             <Text style={userDetailStyles.propertyTitle}>E-mail:</Text>{' '}
             {user?.email}
           </Text>
+
           <Text>
             <Text style={userDetailStyles.propertyTitle}>
               Data de nascimento:
             </Text>{' '}
             {formatDate(user?.birthDate ?? '')}
           </Text>
+
           <Text>
             <Text style={userDetailStyles.propertyTitle}>Permissão:</Text>{' '}
             {roleLabels[user?.role ?? 'default']}
           </Text>
+
           <Text>
             <Text style={userDetailStyles.propertyTitle}>Telefone:</Text>{' '}
             {formatPhoneNumber(user?.phone ?? '')}
